+++ R/download-ao.R
@@ -11,7 +11,9 @@
 #' }
 
 #' @examples
+#' \dontrun{
 #' nao <- download_nao()
+#' }
 #'
 #' @references \url{https://www.ncdc.noaa.gov/teleconnections/nao}
 
@@ -26,7 +28,9 @@ download_ao <- function(){
   
   ao_link ="https://www.ncdc.noaa.gov/teleconnections/ao/data.csv"
   
-  ao = read.csv(ao_link, 
+  res = check_response(ao_link)
+  
+  ao = read.csv(res, 
                    col.names = c("Date","AO"),
                    skip = 1,
                    stringsAsFactors = FALSE)