+++ R/download-nao.R
@@ -11,7 +11,9 @@
 #' }
 
 #' @examples
+#' \dontrun{
 #' nao <- download_nao()
+#' }
 #'
 #' @references \url{https://www.ncdc.noaa.gov/teleconnections/nao}
 
@@ -26,7 +28,9 @@ download_nao <- function(){
   
   nao_link ="https://www.ncdc.noaa.gov/teleconnections/nao/data.csv"
   
-  nao = read.csv(nao_link, 
+  res <- check_response(nao_link)
+  
+  nao = read.csv(res, 
                    col.names = c("Date","NAO"),
                    skip = 1,
                    stringsAsFactors = FALSE)