+++ R/download-npgo.R
@@ -13,7 +13,9 @@
 #' }
 
 #' @examples
+#' \dontrun{
 #' npgo <- download_npgo()
+#' }
 #'
 #' @references \url{http://www.oces.us/npgo} 
 
@@ -24,7 +26,11 @@ download_npgo <- function() {
     return(message("A working internet connection is required to download and import the climate indices."))
   }
   
-  npgo <- read.table("http://www.oces.us/npgo/data/NPGO.txt", 
+  npgo_link ="http://www.oces.us/npgo/data/NPGO.txt"
+  
+  res = check_response(npgo_link)
+  
+  npgo <- read.table(res, 
                      comment.char = "#",
                      col.names = c("Year","Month","NPGO"),
                      stringsAsFactors = FALSE)