+++ R/utils.R
@@ -10,19 +10,21 @@ abbr_month <- function(date){
 }
 
 ## Check the response from server.
-#' @source https://github.com/ropensci/rnoaa/blob/f65d92420e9bb6f4b8d09cbea32c0132638843ce/R/zzz.r#L89-L94
-check_response <- function(response){
-  browser()
+check_response <- function(link){
+  #browser()
   
-  if(response$status_code == 200){
+  response <- curl::curl_fetch_memory(link)
+  
+  if(!response$status_code == 200){
     stop(paste0("Non successful http request. Target server returning a ", response$status_code, " error code"),
         call. = FALSE)
   }
   
   if(grepl("shutdown", response$url)){
-    stop("Data source is currently down due to a government shutdown",
+    stop("Data source is currently unavailable due to a US government shutdown",
          call. = FALSE)
   }
   
-  invisible(TRUE)
+  
+  textConnection(rawToChar(response$content))
 }